pyproject.toml
[tool.poetry]
name = "fast-api"
version = "0.1.0"
description = "Primeira aplicação com o FastAPI"
authors = ["Rafael Martini Silva <rafaelmartinisilva@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.*"
fastapi = "^0.111.0"
pydantic = {extras = ["email"], version = "^2.7.4"}
sqlalchemy = "^2.0.31"
pydantic-settings = "^2.3.4"
pwdlib = {extras = ["argon2"], version = "^0.2.0"}
python-multipart = "^0.0.9"
pyjwt = "^2.8.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.0"
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"
factory-boy = "^3.3.0"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'
run = 'fastapi dev fast_api/app.py --port 8500'
pre_test = 'task lint'
test = 'pytest -s -x --cov=fast_api -vv'  # -x -> para no primeiro teste com falha
# test = 'pytest -s --cov=fast_api -vv'
post_test = 'coverage html'


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
